% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{annotate_peaks}
\alias{annotate_peaks}
\title{annotate peaks}
\usage{
annotate_peaks(
  feature_file,
  peaks_file,
  outfile = FALSE,
  outfileName = "sample"
)
}
\arguments{
\item{feature_file}{A character vector containing path to the file. Either a
bed file, file format should be standard UCSC bed format with column:
\code{ 1. chromosome-id, 2. start, 3. end, 4. name, 5. score, 6. strand} OR
gene feature file with extention \code{.gff or .gff3 }}

\item{peaks_file}{A character vector containing path to the peaks file.
peaks_file can be either \code {1. macs2 callpeak output (i.e. .narrowPeak
or .broadPeak) 2. _peaks.txt, obtained by Homer, analyzeChIP-Seq.pl 3.
_TotalPeaks.tab, obtained by ChIPFun::combine_macs_homer_peaks()}
For macs, 7th column from macs2 output is referred as fold_change. For Homer output,
first 34 lines containing textual information about peaks will be ignored
and 'Clonal Fold Change' column is referred as fold_change.}

\item{outfile}{Logical, to write output in excel sheet or not.}

\item{outfileName}{A character vector to determine name of output file.}
}
\value{
A tibble or output file and pie-chart with target information and classification of
  targets with respect to peak location. Targets are classified as \itemize{
  \item InPromoter: If the peak lies at 5'-upstream of gene start as shown: (^
  ----->), where ^ :peak and -----> :gene in 5'-3' direction.
  \item WithInGeneBody: If the peak completely overlaps with the gene-body region
  (--^-->)
  \item  PartialOverlap_Gene: If the peak partially overlaps at gene
  start (-^---->). Further they are classified as
  PartialOverlap_Gene_1,2,3... if they overlap partially with more than one
  gene, usually the case in compact fungal genomes with many genes
  overlapping either on same strand or different strand.(<---^-
  ----->)
  \item AtGeneEnd: If the peak overlaps with 3'-end of a gene, and there is no gene
  in 5'-3' nearby (----^->)
  \item NearestTarget: Peaks with nearest gene
  promoter at distance > 2000, then assign them to nearest gene irrespective
  of strand. (--A--> ^     <--B--           --C-->  ). Here, ie. even-if the
  peak is in promoter according to gene-C but in NearestTarget category, peak
  will be assigned to gene-A due on its nearest distance }
}
\description{
assign each peak to target genes based on their location
}
\examples{
\dontrun{

bed_file <- "A_nidulans_FGSC_A4_version_s10-m04-r07_cds.bed"
peaks_file <- "myc_narrow_TotalPeaks.tab"
annotate_peaks(feature_file=bed_file,peaks_file=peaks_file, outfile=TRUE,outfileName="myc_peaks")

}
}
