\name{choose_k}
\docType{methods}
\alias{choose_k-method}
\alias{choose_k}
\alias{choose_k,GRanges-method}

\title{
Choice of the final classification of peaks
}
\description{
Selection of the final classification of the peaks, given the desired
number of clusters and the presence or absence of the alignment procedure chosen. 
This choice is usually
driven by the graph returned by \link[=cluster_peak-method]{cluster_peak}.
}
\usage{
\S4method{choose_k}{GRanges}(object, k = NULL, shift.peak = NULL, cleaning = TRUE)
}
\arguments{
  \item{object}{ \linkS4class{GRanges} object. It must contain the metadata columns
  associated to the chosen classification. Further details are provided in \code{shift.peak}.
  }

  \item{k}{integer. Number of chosen clusters.}

  \item{shift.peak}{logical. If \code{TRUE}, the clustering
  with alignment is chosen, if \code{FALSE}, the classification
  without alignment is selected. If \code{shift.peak = TRUE}, \code{object}
  must contain the metadata columns \code{labels_shift},
  \code{coef_shift}, \code{dist_shift}; if \code{shift.peak = FALSE}
  it must contain the columns \code{labels_NOshift}
  and \code{dist_NOshift}.}
  
  \item{cleaning}{logical. If \code{TRUE}, all metadata columns generated by \link{FunChIP} on 
  the \linkS4class{GRanges} object are removed, and one new column  containing 
  the classification result is added: \code{cluster}.
  If \code{FALSE}, the metadata columns generated by \link{FunChIP} are kept.
  Default is \code{TRUE}. See Value for further details on the metadata column added.}

}
\details{
The choice of the optimal number of clusters and the presence of alignment can be guided by
the graph plotted in the \link[=cluster_peak-method]{cluster_peak} method. In particular, for the
optimal number of clusters \eqn{k} the distance 
significantly decreases with respect to the lower values of \eqn{k}, and negligibly increases 
with respect to higher values of \eqn{k} (elbow in the line).
The introduction of the alignment leads to better clustering of the data if
if the global distance is significantly lowered.
registration.
}

\value{
if \code{cleaning = FALSE},
the \linkS4class{GRanges} \code{object} with a new metadata column:
\itemize{
\item{\code{cluster} integer. The label
correspondent to the classification chosen}
}
If \code{cleaning = TRUE}, all the metadata columns added through all the 
analysis are removed and \code{object} is returned with just the metadata column
\code{cluster}
}

\author{
Alice Parodi, Marco J. Morelli, Laura M. Sangalli, Piercesare Secchi, Simone Vantini
}

\seealso{
\link[=cluster_peak-method]{cluster_peak}
}
\examples{

# load the data
data(peaks)

# k = 3 clusters with the alignment
# with integer shifts are chosen

peaks.classified <- choose_k(peaks.data.cluster, k = 3, 
shift.peak = TRUE, cleaning = FALSE)


}

