% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_macs_homer_peaks.R
\name{combine_macs_homer_peaks}
\alias{combine_macs_homer_peaks}
\title{combine peaks from homer and macs2 output}
\usage{
combine_macs_homer_peaks(macs_input, homer_input)
}
\arguments{
\item{macs_input}{A character vector, determining path to MACS2 output. Only
\code{ .narrowPeak or .broadPeak  files obtained by macs2 callpeak }
should be used, with standard MACS2 output
(https://github.com/taoliu/MACS). 7th column from macs2 output is referred as fold_change.}

\item{homer_input}{A character vector, determining path to HOMER output. File
ending with \code{peaks.txt, obtained by analyzeChIP-Seq.pl} should be
used. First 34 lines containing textual information about peaks will be
ignored. 'Clonal Fold Change' column is referred as fold_change}
}
\value{
A tibble, consisting combined targets from MACS2 and HOMER output,
  with their classification as unique to MACS/HOMER or common from both
  output.Peaks are designated common if they have either partial or complete
  overlap.
  \code{.bed} file of combined peaks to upload in genome browser and \code{_TotalPeaks.tab} file to input in \code{annotate_peaks()}
}
\description{
macs2 callpeak and homer: analyzeChIP-Seq.pl output peaks are combined based on overlaps in genomic regions
}
\examples{
\dontrun{

homer_file <- "ABC_HOMER_peaks.txt"
macs_file <- "ABC_narrow_peaks.narrowPeak"
combine_macs_homer_peaks(macs_input=macs_file, homer_input= homer_file)
}

}
