\name{FunChIP-package}
\alias{FunChIP-package}
\alias{FunChIP}
\docType{package}
\title{Clustering and Alignment of ChIP-Seq peaks
based on their shapes}

\description{ 
Efficient implementation of the 
k-mean alignment algorithm to classify spline-smoothed ChIP-Seq peaks.}


\details{
\tabular{ll}{
Package: \tab FunChIP \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2016-04-18 \cr
License: \tab GPL-3\cr
}
}

\author{
Alice Parodi, Marco J. Morelli, Laura M. Sangalli, Piercesare Secchi, Simone Vantini
}

\references{
Sangalli, L.M., Secchi, P., Vantini, S., Vitelli, V., 2010. \emph{"K-mean alignment for curve clustering"}. Computational Statistics and Data Analysis, 54, 1219-1233.}

\seealso{
  \code{\link[=pileup_peak-method]{pileup_peak}, \link[=smooth_peak-method]{smooth_peak},
  \link[=summit_peak-method]{summit_peak}, \link[=distance_peak]{distance_peak},
  \link[=cluster_peak-method]{cluster_peak},
  \link[=choose_k-method]{choose_k}, \link[=plot_peak-method]{plot_peak}}
}


