% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genelist_specific_profileplot.R
\name{genelist_specific_profileplot}
\alias{genelist_specific_profileplot}
\title{profile plot specific to a genelist}
\usage{
genelist_specific_profileplot(
  feature_file,
  genelist,
  bw_files_dir,
  promoter = TRUE,
  bw_pattern = NULL,
  max_scale_limit = 60,
  output = TRUE,
  output_name = "Sample"
)
}
\arguments{
\item{feature_file}{A character vector containing path to the file. Either a
bed file, file format should be standard UCSC bed format with column:
\code{ 1. chromosome-id, 2. start, 3. end, 4. name, 5. score, 6. strand} OR
gene feature file with extention \code{.gff or .gff3 }}

\item{genelist}{A list of genes to be subsetted from the feature file}

\item{bw_files_dir}{A character vector containing path to files \code{.bw or
.bdg}}

\item{promoter}{Logical, if TRUE plot 1kb up/downstream of start co-ordinate;
if FALSE plot 1kb up/downstream of gene-body \code{default: TRUE}}

\item{bw_pattern}{A character vector, to replace a pattern from bw or bdg
file while plotting. Useful when file names are extremely big.
\code{default: NULL}}

\item{max_scale_limit}{Numeric, maximum limit of the color key. \code{default:60}}

\item{output}{Logical, to return a output of profile plot or not.
\code{default: TRUE, with output_name,if FALSE; plot will be returned} **
Highly recommended to plot in outfile when .bw files are more than 2.}

\item{output_name}{A character vector containing name of the output plot.\code{default:Sample}}
}
\value{
Parallel plots of profiles for given gene lists
}
\description{
Profile plot for list of genes to be subsetted from genome feature file.
}
\examples{
\dontrun{

feature_file <- "A_nidulans_FGSC_A4_version_s10-m04-r07_cds.bed"
gene_list <- clipr::read_clip()
bw_files_dir <- "/Users/Pooja/Documents/Data-Analysis/Others/FangFang/2019/ChIP_Analysis"
pattern <- "*_normalized.bw"
genelist_specific_profileplot(feature_file, genelist = gene_list,bw_files_dir = bw_files_dir, promoter = FALSE,bw_pattern = pattern)
}

}
